+++ packages/core/src/components/search/Suggestions/Suggestions.tsx
@@ -1,3 +1,4 @@
+import { useSession } from '@faststore/sdk'
 import { List as UIList } from '@faststore/ui'
 import { gql } from '@vtex/graphql-utils'
 import { useEffect, useState } from 'react'
@@ -14,8 +15,11 @@ import type {
 import SuggestionProductCard from '../SuggestionProductCard'
 
 const SearchSuggestionsQuery = gql`
-  query SearchSuggestionsQuery($term: String!) {
-    search(first: 10, term: $term) {
+  query SearchSuggestionsQuery(
+    $term: String!
+    $selectedFacets: [IStoreSelectedFacet!]
+  ) {
+    search(first: 10, term: $term, selectedFacets: $selectedFacets) {
       suggestions {
         terms
         products {
@@ -87,6 +91,7 @@ export interface SuggestionsProps extends HTMLAttributes<HTMLDivElement> {
 }
 
 function useSuggestions(term: string) {
+  const { channel, locale } = useSession()
   const [suggestions, setSuggestions] =
     useState<SearchSuggestionsQueryQuery['search']['suggestions']>()
 
@@ -98,13 +103,19 @@ function useSuggestions(term: string) {
       request<
         SearchSuggestionsQueryQuery,
         SearchSuggestionsQueryQueryVariables
-      >(SearchSuggestionsQuery, { term })
+      >(SearchSuggestionsQuery, {
+        term,
+        selectedFacets: [
+          { key: 'channel', value: channel ?? '' },
+          { key: 'locale', value: locale },
+        ],
+      })
         .then((data) => {
           setSuggestions(data.search.suggestions)
         })
         .finally(() => setLoading(false))
     }
-  }, [term])
+  }, [term, channel, locale])
 
   const terms = suggestions?.terms ?? []
   const products = suggestions?.products ?? []