+++ packages/core/src/pages/_app.tsx
@@ -33,7 +33,12 @@ function App({ Component, pageProps }: AppProps) {
         actions={uiActions}
         effects={uiEffects}
       >
-        <SessionProvider initialState={{ channel: storeConfig.channel }}>
+        <SessionProvider
+          initialState={{
+            channel: storeConfig.channel,
+            locale: storeConfig.locale,
+          }}
+        >
           <CartProvider mode="optimistic" onValidateCart={validateCart}>
             <ModalProvider>
               <Layout>